포인트 비율 소수점 반영 및 포인트 적립내역 하단부 로직 수정 +관리자페이지 페이징 일부

+++ src/main/java/com/tpay/domains/order/domain/OrderEntity.java
@@ -127,10 +127,9 @@ public class OrderEntity extends BaseTimeEntity {
         return points;
     }
 
-    public long getPointsWithPercentage(Integer balancePercentage){
+    public long getPointsWithPercentage(Double balancePercentage){
         double vat = Double.parseDouble(this.totalVat);
         return (long) Math.floor((vat * balancePercentage) / 100);
-
     }
 
     public OrderEntity setOrderNumber(String orderNumber) {
